requirements.txt
python-telegram-bot
telethon
python-dotenv
black
pytest
pylint
aiosqlite